Create README.md
An experimental sorting algorithm that uses a tree with N leaves. Each node copies its lesser child and then propagates the evaluation to it's parent to ensure the root node contains the index of the next minimum value. After an element is sorted, the corresponding leaf is closed and it's parent reevaluates.